feat(delay): 100% completed
