--- src/audio_effect/aa_delay_effect.h
@@ -1,10 +1,0 @@
-//
-// Created by william on 2020/2/21.
-//
-
-#pragma once
-
-namespace libaa
-{
-
-}
